Initial add of geoparser files
Initial add of some of the geoparsers used in JournalMap. Note that the
location_parser.py is an older, Python version from what is currently
being used in JournalMap. This older version was a modification of the
GeoLucidate code. The newer version was written from scratch and is
both more flexible and capable. The bounding box parser is not yet
implemented in JournalMap but is a test script.